[sui-core] Extract transaction simulation into its own module
