+++ crates/sui-core/src/authority/transaction_simulation.rs
@@ -1,0 +1,279 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+use super::*;
+
+pub(super) fn simulate_transaction(
+    state: &AuthorityState,
+    mut transaction: TransactionData,
+    checks: TransactionChecks,
+    allow_mock_gas_coin: bool,
+) -> SuiResult<SimulateTransactionResult> {
+    if transaction.kind().is_system_tx() {
+        return Err(SuiErrorKind::UnsupportedFeatureError {
+            error: "simulate does not support system transactions".to_string(),
+        }
+        .into());
+    }
+
+    let epoch_store = state.load_epoch_store_one_call_per_task();
+    if !state.is_fullnode(&epoch_store) {
+        return Err(SuiErrorKind::UnsupportedFeatureError {
+            error: "simulate is only supported on fullnodes".to_string(),
+        }
+        .into());
+    }
+
+    let dev_inspect = checks.disabled();
+    if dev_inspect && state.config.dev_inspect_disabled {
+        return Err(SuiErrorKind::UnsupportedFeatureError {
+            error: "simulate with checks disabled is not allowed on this node".to_string(),
+        }
+        .into());
+    }
+
+    // Reject coin reservations in gas payment when the execution engine
+    // doesn't support them.
+    let protocol_config = epoch_store.protocol_config();
+    if !protocol_config.enable_coin_reservation_obj_refs()
+        && transaction.gas().iter().any(|obj_ref| {
+            sui_types::coin_reservation::ParsedDigest::is_coin_reservation_digest(&obj_ref.2)
+        })
+    {
+        return Err(SuiErrorKind::UnsupportedFeatureError {
+            error: "coin reservations in gas payment are not supported at this protocol version"
+                .to_string(),
+        }
+        .into());
+    }
+
+    // Compute input/receiving object kinds before mock gas injection so the mock
+    // gas reference is not included in input_object_kinds (it is added to
+    // input_objects directly after object loading).
+    let input_object_kinds = transaction.input_objects()?;
+    let receiving_object_refs = transaction.receiving_objects();
+
+    // Inject mock gas coin before validity_check so that on protocol versions
+    // where address-balance gas payments are not yet enabled, the non-empty
+    // payment check in validity_check passes for simulate/dev-inspect requests
+    // submitted without explicit gas.
+    // Also required before pre_object_load_checks so that funds-withdrawal
+    // processing sees non-empty payment and doesn't create an address-balance
+    // withdrawal for gas.
+    // Skip mock gas for gasless transactions — they don't use gas coins.
+    let is_gasless = protocol_config.enable_gasless() && transaction.is_gasless_transaction();
+    let mock_gas_object = if allow_mock_gas_coin && transaction.gas().is_empty() && !is_gasless {
+        let obj = Object::new_move(
+            MoveObject::new_gas_coin(
+                OBJECT_START_VERSION,
+                ObjectID::MAX,
+                DEV_INSPECT_GAS_COIN_VALUE,
+            ),
+            Owner::AddressOwner(transaction.gas_data().owner),
+            TransactionDigest::genesis_marker(),
+        );
+        transaction.gas_data_mut().payment = vec![obj.compute_object_reference()];
+        Some(obj)
+    } else {
+        None
+    };
+
+    // Full validity check including gas budget and price.
+    transaction.validity_check(&epoch_store.tx_validity_check_context())?;
+
+    let declared_withdrawals = state.pre_object_load_checks(
+        &transaction,
+        &[],
+        &input_object_kinds,
+        &receiving_object_refs,
+        epoch_store.protocol_config(),
+    )?;
+    let address_funds: BTreeSet<_> = declared_withdrawals.keys().cloned().collect();
+
+    let (mut input_objects, receiving_objects) = state.input_loader.read_objects_for_signing(
+        // We don't want to cache this transaction since it's a simulation.
+        None,
+        &input_object_kinds,
+        &receiving_object_refs,
+        epoch_store.epoch(),
+    )?;
+
+    // Add mock gas to input objects after loading (it doesn't exist in the store).
+    let mock_gas_id = mock_gas_object.map(|obj| {
+        let id = obj.id();
+        input_objects.push(ObjectReadResult::new_from_gas_object(&obj));
+        id
+    });
+
+    let protocol_config = epoch_store.protocol_config();
+
+    let (gas_status, checked_input_objects) = if dev_inspect {
+        sui_transaction_checks::check_dev_inspect_input(
+            protocol_config,
+            &transaction,
+            input_objects,
+            receiving_objects,
+            epoch_store.reference_gas_price(),
+        )?
+    } else {
+        sui_transaction_checks::check_transaction_input(
+            epoch_store.protocol_config(),
+            epoch_store.reference_gas_price(),
+            &transaction,
+            input_objects,
+            &receiving_objects,
+            &state.metrics.bytecode_verifier_metrics,
+            &state.config.verifier_signing_config,
+        )?
+    };
+
+    let executor = epoch_store.simulate_executor();
+
+    let (mut kind, signer, gas_data) = transaction.execution_parts();
+    let rewritten_inputs = rewrite_transaction_for_coin_reservations(
+        state.chain_identifier,
+        &*state.coin_reservation_resolver,
+        signer,
+        &mut kind,
+        None,
+    )?;
+    let early_execution_error = get_early_execution_error(
+        &transaction.digest(),
+        &checked_input_objects,
+        state.config.certificate_deny_config.certificate_deny_set(),
+        &FundsWithdrawStatus::MaybeSufficient,
+    );
+    // Dev-inspect/simulation path (not committed): no assigned accumulator version here, so the
+    // IFFW short-circuit applies unconditionally (`None`), matching non-mainnet execution.
+    let execution_params = match early_execution_error {
+        None => ExecutionOrEarlyError::ok(None),
+        Some(errors) => ExecutionOrEarlyError::failed(errors, None),
+    };
+
+    let tracking_store = TrackingBackingStore::new(state.get_backing_store().as_ref());
+
+    // Clone inputs for potential retry if object funds check fails post-execution.
+    let cloned_input_objects = checked_input_objects.clone();
+    let cloned_gas = gas_data.clone();
+    let cloned_kind = kind.clone();
+    let tx_digest = transaction.digest();
+    let epoch_id = epoch_store.epoch_start_config().epoch_data().epoch_id();
+    let epoch_timestamp_ms = epoch_store
+        .epoch_start_config()
+        .epoch_data()
+        .epoch_start_timestamp();
+    let (inner_temp_store, _, effects, execution_result) = executor.dev_inspect_transaction(
+        &tracking_store,
+        protocol_config,
+        state.metrics.execution_metrics.clone(),
+        false, // expensive_checks
+        execution_params,
+        &epoch_id,
+        epoch_timestamp_ms,
+        checked_input_objects,
+        gas_data,
+        gas_status,
+        kind,
+        rewritten_inputs.clone(),
+        signer,
+        tx_digest,
+        dev_inspect,
+    );
+
+    // Post-execution: check object funds (non-address withdrawals discovered during execution).
+    let (inner_temp_store, effects, execution_result) = if execution_result.is_ok() {
+        let has_insufficient_object_funds = inner_temp_store
+            .accumulator_running_max_withdraws
+            .iter()
+            .filter(|(id, _)| !address_funds.contains(id))
+            .any(|(id, max_withdraw)| {
+                let balance = state.get_account_funds_read().get_latest_account_amount(id);
+                balance < *max_withdraw
+            });
+
+        if has_insufficient_object_funds {
+            let retry_gas_status = SuiGasStatus::new(
+                cloned_gas.budget,
+                cloned_gas.price,
+                epoch_store.reference_gas_price(),
+                protocol_config,
+            )?;
+            let (store, _, effects, result) = executor.dev_inspect_transaction(
+                &tracking_store,
+                protocol_config,
+                state.metrics.execution_metrics.clone(),
+                false,
+                ExecutionOrEarlyError::failed(
+                    NonEmpty::new(ExecutionErrorKind::InsufficientFundsForWithdraw),
+                    None,
+                ),
+                &epoch_id,
+                epoch_timestamp_ms,
+                cloned_input_objects,
+                cloned_gas,
+                retry_gas_status,
+                cloned_kind,
+                rewritten_inputs,
+                signer,
+                tx_digest,
+                dev_inspect,
+            );
+            (store, effects, result)
+        } else {
+            (inner_temp_store, effects, execution_result)
+        }
+    } else {
+        (inner_temp_store, effects, execution_result)
+    };
+
+    let loaded_runtime_objects = tracking_store.into_read_objects();
+    let unchanged_loaded_runtime_objects =
+        crate::transaction_outputs::unchanged_loaded_runtime_objects(
+            &transaction,
+            &effects,
+            &loaded_runtime_objects,
+        );
+
+    let object_set = {
+        let objects = {
+            let mut objects = loaded_runtime_objects;
+
+            for o in inner_temp_store
+                .input_objects
+                .into_values()
+                .chain(inner_temp_store.written.into_values())
+            {
+                objects.insert(o);
+            }
+
+            objects
+        };
+
+        let object_keys = sui_types::storage::get_transaction_object_set(
+            &transaction,
+            &effects,
+            &unchanged_loaded_runtime_objects,
+        );
+
+        let mut set = sui_types::full_checkpoint_content::ObjectSet::default();
+        for k in object_keys {
+            if let Some(o) = objects.get(&k) {
+                set.insert(o.clone());
+            }
+        }
+
+        set
+    };
+
+    Ok(SimulateTransactionResult {
+        objects: object_set,
+        events: effects.events_digest().map(|_| inner_temp_store.events),
+        effects,
+        execution_result,
+        mock_gas_id,
+        unchanged_loaded_runtime_objects,
+        suggested_gas_price: state
+            .congestion_tracker
+            .get_suggested_gas_prices(&transaction),
+    })
+}
